Get matching episode info
For now, print it out to the screen in formatted json, and collect the relevant values for constructing a filename. Also fix user input so that a blank string is not accepted. This solves the problem where a blank search returns all episodes, because python evaluates `'' in 'a string of words'` to True.