Print out capitalized titles with punctuation
Use a separate list of unmodified episode titles and transform them during iteration when they need to match the keys in the episode_names_ids dictionary. Create a separate function to remove punctuation, so that we can do it outside of main() when we need to. Make the list of series names and episode titles both tuples, because they should be immutable.