Add warnings when you pass a file to --symlinks
If you leave off the argument to the -l option, and you type -l right before the positional arguments (episode files), you'll accidentally pass the first episode file to -l. This situation was not dangerous (it just gets ignored), but it is confusing. Now there are warnings when this happens.