Handle errors when renaming files
If the os can't create a symlink, print an error and exit with a bookmark.
Do the same if filenames contain characters prohibited by the OS. Remove
question mark from default 'allowed_chars' in the config. Add 'status'
argument to quit() function so that you don't always have to exit with
status code 0.